LayoutTests/platform/ios-simulator/editing/inserting/insert-composition-whitespace-expected.txt

CONSOLE MESSAGE: line 31: ReferenceError: Can't find variable: textInputController
FAIL successfullyParsed should be true (of type boolean). Was undefined (of type undefined).

TEST COMPLETE
